send key event with socket #5
