+++ package.json
@@ -4,15 +4,15 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "node ./src/app.mjs",
-    "dev": "nodemon ./src/app.mjs"
+    "start": "node ./src/app.js",
+    "dev": "nodemon ./src/app.js"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "express": "^4.18.2",
-    "socket.io": "^4.7.4"
+    "express": "4.14.0",
+    "socket.io": "1.4.8"
   },
   "devDependencies": {
     "nodemon": "^3.0.3"
