+++ src/client/client.js
@@ -2,12 +2,12 @@ const socket = io();
 
 // Create a 10x20 game board
 for (let i = 0; i < 200; i++) {
-    let child = document.createElement('li');
-    board.appendChild(child);
+  let child = document.createElement('li');
+  board.appendChild(child);
 }
 
 // Handle arrow key press and send move event to server
-document.addEventListener('keydown', (event) => {
+document.addEventListener('keydown', event => {
   let direction = null;
   switch (event.key) {
     case 'ArrowUp':
@@ -22,15 +22,24 @@ document.addEventListener('keydown', (event) => {
     case 'ArrowRight':
       direction = 'right';
       break;
+    case ' ':
+      direction = 'sprint';
+      break;
   }
 
   if (direction) {
-    socket.emit('move', { direction });
+    socket.emit('move', { type: 'keydown', direction: direction });
   }
 });
 
-// Listen for game state updates from the server
-socket.on('update', (data) => {
+document.addEventListener('keyup', event => {
+  if (event.key === 'ArrowDown') {
+    socket.emit('move', { type: 'keyup', direction: 'down' });
+  }
+});
+
+socket.on('update', data => {
   const gameState = data.gameState;
+  console.log('Received game state:', gameState);
   // Update the client-side game interface based on the received game state
 });
