+++ src/client/index.html
@@ -1,36 +1,21 @@
-<!doctype html>
-<html lang="en">
+<!DOCTYPE html>
+<html>
   <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script defer src="client.js"></script>
+    <meta charset="utf-8" />
+    <title>Socket.io Chat Example</title>
+    <link
+      rel="stylesheet"
+      href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"
+    />
     <link rel="stylesheet" href="style.css" />
-    <title>Red Tetris</title>
   </head>
   <body>
-    <h1>red-tetris</h1>
-    <ul id="board"></ul>
-    <button onclick="newGame()">new game</button>
-    <button onclick="stopGame()">stop game</button>
+    <div class="container">
+      <h1>red-tetris</h1>
+      <ul id="board"></ul>
+    </div>
+    <script src="/socket.io/socket.io.js"></script>
+    <!-- <script src="test.js"></script> -->
+    <script src="client.js"></script>
   </body>
-  <script>
-    init();
-function init() {
-  for (let i = 0; i < 200; i++) {
-    let child = document.createElement('li');
-    board.appendChild(child);
-  }
-  document.addEventListener('keydown', e => {
-    console.log('[debug] down');
-    if (movements.hasOwnProperty(e.key)) {
-      movements[e.key](fallingPiece);
-    }
-  });
-  document.addEventListener('keyup', e => {
-    if (e.key === 'ArrowDown') {
-      resetSpeed(fallingPiece);
-    }
-  });
-}
-  </script>
 </html>
